- introduce common compose module with common helpers and extensions - introduce new dimensions configuration object to offer more flexibility for the UI (this is a breaking API change) - add snapshot tests for m2 and m3 alike - simplify gradle setup for m2 / m3 modules - remove fake data from m2/m3 and only have in snapshot test
