+++ aboutlibraries-compose-m2/src/commonMain/kotlin/com/mikepenz/aboutlibraries/ui/compose/Libraries.kt
@@ -6,15 +6,10 @@ import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-import androidx.compose.runtime.produceState
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.mikepenz.aboutlibraries.Libs
 import com.mikepenz.aboutlibraries.entity.Library
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 
 /**
  * Displays all provided libraries in a simple list.
@@ -31,8 +26,7 @@ fun LibrariesContainer(
     showLicenseBadges: Boolean = true,
     colors: LibraryColors = LibraryDefaults.libraryColors(),
     padding: LibraryPadding = LibraryDefaults.libraryPadding(),
-    itemContentPadding: PaddingValues = LibraryDefaults.ContentPadding,
-    itemSpacing: Dp = LibraryDefaults.LibraryItemSpacing,
+    dimensions: LibraryDimensions = LibraryDefaults.libraryDimensions(),
     header: (LazyListScope.() -> Unit)? = null,
     onLibraryClick: ((Library) -> Unit)? = null,
 ) {
@@ -48,8 +42,7 @@ fun LibrariesContainer(
         showLicenseBadges = showLicenseBadges,
         colors = colors,
         padding = padding,
-        itemContentPadding = itemContentPadding,
-        itemSpacing = itemSpacing,
+        dimensions = dimensions,
         header = header,
         onLibraryClick = onLibraryClick,
         licenseDialogBody = { library ->
@@ -72,8 +65,7 @@ fun LibrariesContainer(
     showLicenseBadges: Boolean = true,
     colors: LibraryColors = LibraryDefaults.libraryColors(),
     padding: LibraryPadding = LibraryDefaults.libraryPadding(),
-    itemContentPadding: PaddingValues = LibraryDefaults.ContentPadding,
-    itemSpacing: Dp = LibraryDefaults.LibraryItemSpacing,
+    dimensions: LibraryDimensions = LibraryDefaults.libraryDimensions(),
     header: (LazyListScope.() -> Unit)? = null,
     onLibraryClick: ((Library) -> Unit)? = null,
 ) {
@@ -90,40 +82,11 @@ fun LibrariesContainer(
         showLicenseBadges = showLicenseBadges,
         colors = colors,
         padding = padding,
-        itemContentPadding = itemContentPadding,
-        itemSpacing = itemSpacing,
+        dimensions = dimensions,
         header = header,
         onLibraryClick = onLibraryClick,
         licenseDialogBody = { library ->
             Text(library.licenses.firstOrNull()?.licenseContent ?: "")
         }
     )
-}
-
-/**
- * Creates a State<Libs?> that holds the [Libs] as loaded by the [libraries].
- *
- * @see Libs
- */
-@Composable
-fun rememberLibraries(
-    libraries: ByteArray,
-): State<Libs?> = rememberLibraries {
-    libraries.decodeToString()
-}
-
-/**
- * Creates a State<Libs?> that holds the [Libs] as loaded by the [block].
- *
- * @see Libs
- */
-@Composable
-fun rememberLibraries(
-    block: suspend () -> String,
-): State<Libs?> {
-    return produceState<Libs?>(initialValue = null) {
-        value = withContext(Dispatchers.Default) {
-            Libs.Builder().withJson(block()).build()
-        }
-    }
 }
