+++ aboutlibraries-compose-m3/src/commonMain/kotlin/com/mikepenz/aboutlibraries/ui/compose/m3/SharedLibraries.kt
@@ -27,8 +27,6 @@ import androidx.compose.material3.TextButton
 import androidx.compose.material3.Typography
 import androidx.compose.material3.contentColorFor
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
-import androidx.compose.runtime.Stable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -37,14 +35,18 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import com.mikepenz.aboutlibraries.Libs
 import com.mikepenz.aboutlibraries.entity.Library
-import com.mikepenz.aboutlibraries.ui.compose.m3.util.author
-import com.mikepenz.aboutlibraries.ui.compose.m3.util.htmlReadyLicenseContent
+import com.mikepenz.aboutlibraries.ui.compose.DefaultLibraryColors
+import com.mikepenz.aboutlibraries.ui.compose.LibraryColors
+import com.mikepenz.aboutlibraries.ui.compose.LibraryDefaults
+import com.mikepenz.aboutlibraries.ui.compose.LibraryDimensions
+import com.mikepenz.aboutlibraries.ui.compose.LibraryPadding
+import com.mikepenz.aboutlibraries.ui.compose.util.author
+import com.mikepenz.aboutlibraries.ui.compose.util.htmlReadyLicenseContent
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
@@ -64,8 +66,7 @@ fun LibrariesContainer(
     showLicenseBadges: Boolean = true,
     colors: LibraryColors = LibraryDefaults.libraryColors(),
     padding: LibraryPadding = LibraryDefaults.libraryPadding(),
-    itemContentPadding: PaddingValues = LibraryDefaults.ContentPadding,
-    itemSpacing: Dp = LibraryDefaults.LibraryItemSpacing,
+    dimensions: LibraryDimensions = LibraryDefaults.libraryDimensions(),
     header: (LazyListScope.() -> Unit)? = null,
     onLibraryClick: ((Library) -> Unit)? = null,
     licenseDialogBody: (@Composable (Library) -> Unit)? = null,
@@ -87,8 +88,7 @@ fun LibrariesContainer(
         showLicenseBadges = showLicenseBadges,
         colors = colors,
         padding = padding,
-        itemContentPadding = itemContentPadding,
-        itemSpacing = itemSpacing,
+        dimensions = dimensions,
         header = header,
         onLibraryClick = { library ->
             val license = library.licenses.firstOrNull()
@@ -177,8 +177,7 @@ fun Libraries(
     showLicenseBadges: Boolean = true,
     colors: LibraryColors = LibraryDefaults.libraryColors(),
     padding: LibraryPadding = LibraryDefaults.libraryPadding(),
-    itemContentPadding: PaddingValues = LibraryDefaults.ContentPadding,
-    itemSpacing: Dp = LibraryDefaults.LibraryItemSpacing,
+    dimensions: LibraryDimensions = LibraryDefaults.libraryDimensions(),
     header: (LazyListScope.() -> Unit)? = null,
     onLibraryClick: ((Library) -> Unit)? = null,
 ) {
@@ -186,7 +185,7 @@ fun Libraries(
 
     LazyColumn(
         modifier,
-        verticalArrangement = Arrangement.spacedBy(itemSpacing),
+        verticalArrangement = Arrangement.spacedBy(dimensions.libraryItemSpacing),
         state = lazyListState,
         contentPadding = contentPadding
     ) {
@@ -199,7 +198,6 @@ fun Libraries(
             showLicenseBadges = showLicenseBadges,
             colors = colors,
             padding = padding,
-            itemContentPadding = itemContentPadding
         ) { library ->
             val license = library.licenses.firstOrNull()
             if (onLibraryClick != null) {
@@ -225,7 +223,6 @@ internal inline fun LazyListScope.libraryItems(
     showLicenseBadges: Boolean = true,
     colors: LibraryColors,
     padding: LibraryPadding,
-    itemContentPadding: PaddingValues = LibraryDefaults.ContentPadding,
     crossinline onLibraryClick: ((Library) -> Unit),
 ) {
     items(libraries) { library ->
@@ -237,7 +234,6 @@ internal inline fun LazyListScope.libraryItems(
             showLicenseBadges = showLicenseBadges,
             colors = colors,
             padding = padding,
-            contentPadding = itemContentPadding
         ) {
             onLibraryClick.invoke(library)
         }
@@ -246,7 +242,7 @@ internal inline fun LazyListScope.libraryItems(
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalLayoutApi::class)
 @Composable
-internal fun Library(
+fun Library(
     library: Library,
     showAuthor: Boolean = true,
     showDescription: Boolean = false,
@@ -254,7 +250,6 @@ internal fun Library(
     showLicenseBadges: Boolean = true,
     colors: LibraryColors = LibraryDefaults.libraryColors(),
     padding: LibraryPadding = LibraryDefaults.libraryPadding(),
-    contentPadding: PaddingValues = LibraryDefaults.ContentPadding,
     typography: Typography = MaterialTheme.typography,
     onClick: () -> Unit,
 ) {
@@ -263,7 +258,7 @@ internal fun Library(
             .fillMaxWidth()
             .background(colors.backgroundColor)
             .clickable { onClick.invoke() }
-            .padding(contentPadding),
+            .padding(padding.contentPadding),
         verticalArrangement = Arrangement.spacedBy(padding.verticalPadding)
     ) {
         Row(
@@ -326,139 +321,27 @@ internal fun Library(
     }
 }
 
-
 /**
- * Contains the default values used by [Library]
+ * Creates a [LibraryColors] that represents the default colors used in
+ * a [Library].
+ *
+ * @param backgroundColor the background color of this [Library]
+ * @param contentColor the content color of this [Library]
+ * @param badgeBackgroundColor the badge background color of this [Library]
+ * @param badgeContentColor the badge content color of this [Library]
+ * @param dialogConfirmButtonColor the dialog's confirm button color of this [Library]
  */
-object LibraryDefaults {
-    private val LibraryItemPadding = 16.dp
-    private val LibraryNamePaddingTop = 4.dp
-    private val LibraryVersionPaddingStart = 8.dp
-    private val LibraryBadgePaddingTop = 8.dp
-    private val LibraryBadgePaddingEnd = 4.dp
-    internal val LibraryItemSpacing = 0.dp
-
-    /**
-     * The default content padding used by [Library]
-     */
-    val ContentPadding = PaddingValues(LibraryItemPadding)
-
-    /**
-     * Creates a [LibraryColors] that represents the default colors used in
-     * a [Library].
-     *
-     * @param backgroundColor the background color of this [Library]
-     * @param contentColor the content color of this [Library]
-     * @param badgeBackgroundColor the badge background color of this [Library]
-     * @param badgeContentColor the badge content color of this [Library]
-     * @param dialogConfirmButtonColor the dialog's confirm button color of this [Library]
-     */
-    @Composable
-    fun libraryColors(
-        backgroundColor: Color = MaterialTheme.colorScheme.background,
-        contentColor: Color = contentColorFor(backgroundColor),
-        badgeBackgroundColor: Color = MaterialTheme.colorScheme.primary,
-        badgeContentColor: Color = contentColorFor(badgeBackgroundColor),
-        dialogConfirmButtonColor: Color = MaterialTheme.colorScheme.primary,
-    ): LibraryColors = DefaultLibraryColors(
-        backgroundColor = backgroundColor,
-        contentColor = contentColor,
-        badgeBackgroundColor = badgeBackgroundColor,
-        badgeContentColor = badgeContentColor,
-        dialogConfirmButtonColor = dialogConfirmButtonColor,
-    )
-
-    /**
-     * Creates a [LibraryPadding] that represents the default paddings used in a [Library]
-     *
-     * @param namePadding the padding around the name shown as part of a [Library]
-     * @param versionPadding the padding around the version shown as part of a [Library]
-     * @param badgePadding the padding around a badge element shown as part of a [Library]
-     * @param badgeContentPadding the padding around the content of a badge element shown as part of a [Library]
-     * @param verticalPadding the vertical padding between the individual items in the library element
-     */
-    @Composable
-    fun libraryPadding(
-        namePadding: PaddingValues = PaddingValues(top = LibraryNamePaddingTop),
-        versionPadding: PaddingValues = PaddingValues(start = LibraryVersionPaddingStart),
-        badgePadding: PaddingValues = PaddingValues(
-            top = LibraryBadgePaddingTop,
-            end = LibraryBadgePaddingEnd
-        ),
-        badgeContentPadding: PaddingValues = PaddingValues(0.dp),
-        verticalPadding: Dp = 2.dp,
-    ): LibraryPadding = DefaultLibraryPadding(
-        namePadding = namePadding,
-        versionPadding = versionPadding,
-        badgePadding = badgePadding,
-        badgeContentPadding = badgeContentPadding,
-        verticalPadding = verticalPadding,
-    )
-}
-
-/**
- * Represents the background and content colors used in a library.
- */
-@Stable
-interface LibraryColors {
-    /** Represents the background color for this library item. */
-    val backgroundColor: Color
-
-    /** Represents the content color for this library item. */
-    val contentColor: Color
-
-    /** Represents the badge background color for this library item. */
-    val badgeBackgroundColor: Color
-
-    /** Represents the badge content color for this library item. */
-    val badgeContentColor: Color
-
-    /** Represents the text color of the dialog's confirm button  */
-    val dialogConfirmButtonColor: Color
-}
-
-/**
- * Default [LibraryColors].
- */
-@Immutable
-private class DefaultLibraryColors(
-    override val backgroundColor: Color,
-    override val contentColor: Color,
-    override val badgeBackgroundColor: Color,
-    override val badgeContentColor: Color,
-    override val dialogConfirmButtonColor: Color,
-) : LibraryColors
-
-
-/**
- * Represents the padding values used in a library.
- */
-@Stable
-interface LibraryPadding {
-    /** Represents the padding around the name shown as part of a [Library] */
-    val namePadding: PaddingValues
-
-    /** Represents the padding around the version shown as part of a [Library] */
-    val versionPadding: PaddingValues
-
-    /** Represents the padding around a badge element shown as part of a [Library] */
-    val badgePadding: PaddingValues
-
-    /** Represents the padding around the content of a badge element shown as part of a [Library] */
-    val badgeContentPadding: PaddingValues
-
-    /** Represents the vertical padding between the individual items in the library element */
-    val verticalPadding: Dp
-}
-
-/**
- * Default [LibraryPadding].
- */
-@Immutable
-private class DefaultLibraryPadding(
-    override val namePadding: PaddingValues,
-    override val versionPadding: PaddingValues,
-    override val badgePadding: PaddingValues,
-    override val badgeContentPadding: PaddingValues,
-    override val verticalPadding: Dp,
-) : LibraryPadding
+@Composable
+fun LibraryDefaults.libraryColors(
+    backgroundColor: Color = MaterialTheme.colorScheme.background,
+    contentColor: Color = contentColorFor(backgroundColor),
+    badgeBackgroundColor: Color = MaterialTheme.colorScheme.primary,
+    badgeContentColor: Color = contentColorFor(badgeBackgroundColor),
+    dialogConfirmButtonColor: Color = MaterialTheme.colorScheme.primary,
+): LibraryColors = DefaultLibraryColors(
+    backgroundColor = backgroundColor,
+    contentColor = contentColor,
+    badgeBackgroundColor = badgeBackgroundColor,
+    badgeContentColor = badgeContentColor,
+    dialogConfirmButtonColor = dialogConfirmButtonColor,
+)
