+++ aboutlibraries-compose/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=AboutLibraries Compose UI Library
+POM_DESCRIPTION=AboutLibraries automatically detects all dependencies of a project and collects their information including the license. Optionally visualising it via the provided ui components.
+POM_ARTIFACT_ID=aboutlibraries-compose-core
