+++ aboutlibraries-compose/src/commonMain/kotlin/com/mikepenz/aboutlibraries/ui/compose/util/Extensions.kt
@@ -1,10 +1,10 @@
-package com.mikepenz.aboutlibraries.ui.compose.m3.util
+package com.mikepenz.aboutlibraries.ui.compose.util
 
 import com.mikepenz.aboutlibraries.entity.Library
 import com.mikepenz.aboutlibraries.entity.License
 
 val Library.author: String
-    get() = developers.takeIf { it.isNotEmpty() }?.map { it.name }?.joinToString(", ")
+    get() = developers.takeIf { it.isNotEmpty() }?.mapNotNull { it.name }?.joinToString(", ")
         ?: organization?.name ?: ""
 
 val License.htmlReadyLicenseContent: String?
