+++ settings.gradle.kts
@@ -27,6 +27,7 @@ dependencyResolutionManagement {
 
 include(":aboutlibraries-core")
 include(":aboutlibraries")
+include(":aboutlibraries-compose")
 include(":aboutlibraries-compose-m2")
 include(":aboutlibraries-compose-m3")
